Enable headless scraper in initialization
Now the scraper can be run without GUI (TK). TK can be installed without problems as far as I know (should be tested in future). With the param --headless you can completely disable the GUI elements, otherwise it will try launching TK and if it fails it switches over to manual PATH.