Add custom identifier to url
The old structure of reading the csv file and getting the URLs with KVK is now supported again. The identifier is not bound to KVK but may be anything. This identifier is also added to the output tsv or db. This identifier is an addition for the package as users might want to number the urls or add context-specific information that is directly tagged to the urls in the output format which helps keep track of urls throughout the process.